pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["korpaxdev <korpaxdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
fastapi = "^0.104.1"
uvicorn = { extras = ["standart"], version = "^0.24.0.post1" }
pydantic-settings = "^2.1.0"
bcrypt = "^4.0.1"
pydantic = { extras = ["email"], version = "^2.5.1" }
python-jose = "^3.3.0"
pytimeparse = "^1.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
extend-exclude = "database/data/*"
[tool.flake8]
max-line-length = 120
extend-exclude = ".venv"
